Moved test scripts to tests/ dir
